+++ packages/plugin-local-storage/src/services/checks/index.ts
@@ -1,5 +1,0 @@
-import validatePresignedToken from "./validate-presigned-token.js";
-
-export default {
-	validatePresignedToken,
-};