+++ packages/plugin-local-storage/src/types/types.ts
@@ -1,6 +1,0 @@
-export interface PluginOptions {
-	/** The directory where the files will be uploaded */
-	uploadDir: string;
-	/** The secret key used to sign the URLs */
-	secretKey: string;
-}