+++ packages/plugin-s3/src/index.ts
@@ -1,6 +1,6 @@
 import type { Config } from "@lucidcms/core/types";
-import type { PluginOptions } from "./types/types.js";
 import plugin from "./plugin.js";
+import type { PluginOptions } from "./types/types.js";
 
 const lucidS3Plugin = (pluginOptions: PluginOptions) => (config: Config) =>
 	plugin(config, pluginOptions);